rename testReconfiguration to swReconfiguration directory

+++ tests/virtex2pro_test.sh
@@ -30,15 +30,23 @@ function testCase {
     cat > received.txt < /dev/ttyS0&
     PID=$!
     stty 580:5:cbd:8a3b:3:1c:7f:15:4:0:1:0:11:13:1a:0:12:f:17:16:0:0:0:0:0:0:0:0:0:0:0:0:0:0:0:0 -F /dev/ttyS0 #stty -g /dev/ttyS0
+    set +e
     make -f custom.make download >> $oldPWD/work/virtex2pro_output.log
+    if [ $? -ne 0 ]
+    then
+      echo "$1 compilation failed"
+      echo "Log file: work/virtex2pro_output.log"
+      exit 1
+    fi
+    set -e
     sleep 10
     kill $PID
     #wait $PID 2>/dev/null
-    if diff received.txt received_expected.txt >/dev/null ; then
+    if diff -bB received.txt received_expected.txt >/dev/null ; then
       echo "$1 succeeded"
     else
-      echo "$1 failed"
-      echo "Log file: work/virtex2pro_output.log"
+      echo "$1 failed (output mismatch)"
+      echo "Output file: ../examples/$1/received.txt"
       exit 1
     fi
     cd - >/dev/null
